requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
Authlib==1.3.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==4.42
cryptography==41.0.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
dnspython==2.4.2
fastapi==0.109.0
fastavro==1.9.3
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
joblib==1.3.2
llama-index==0.9.30
loguru==0.7.2
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.7.2
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pinecone-client==2.2.4
platformdirs==4.1.0
portalocker==2.8.2
protobuf==4.25.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pypdf==3.17.4
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.7.0
regex==2023.12.25
requests==2.31.0
ruff==0.1.13
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
validators==0.22.0
vulture==2.10
watchfiles==0.21.0
weaviate-client==3.26.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
